Use product network in matrix mult. example.
To me this seems easier to understand because the reader does not need
to figure out what parts are related to the matrix multiplication and
which are just there to calculate products. If the reader wants to know
how the product can be calculated, we have the scalar multiplication
example.

This commit also fixes some problems when changing the matrix
dimensions.